social_media/requirements.txt
fastapi
uvicorn
flask
psycopg2-binary
requests
faker
